🚚 move: MainLayout device 폴더 분리
