+++ app/src/components/layouts/MainLayout/index.tsx
@@ -1,7 +1,7 @@
 import { Desktop, Mobile, Tablet } from '@utils/responsive/Device';
-import { DesktopMainLayout } from './DesktopMainLayout';
-import { MobileMainLayout } from './MobileMainLayout';
-import { TabletMainLayout } from './TabletMainLayout';
+import { DesktopMainLayout } from './Desktop';
+import { MobileMainLayout } from './Mobile';
+import { TabletMainLayout } from './Tablet';
 
 export const MainLayout = () => {
   return (
